To avoid massive ugly Makefile.am changes, created a src_cuda dir, with contains symlinks to all the source files in src/ but renamed .cu (for nvcc). Plus a CUDA-specific Makefile.am which is very different from the one in src, for one thing libtool is NOT used. Changes to configure.ac and top-level Makefile.am were necessary to accommodate this.
Now when --enable-cuda is specified, libtool is not used anywhere. This may be
required?
Now when --enable-cuda is specified, only CC=nvcc works, but it will be honored
in Makefile.am
CUFLAGS moved from Makefile.am files to configure.ac. --enable-debug will do
the "right thing" for the CUDA files as well as CPU files, but they are
nvcc-specific

Bugfix in gptl_device.cu for "how did this ever work" in cuda_acc build
(calling cudaMalloc for a host variable) which failed as one would normally
expect in the new code. And the new code is cleaner.

Massive changes to cuda/tests/Makefile.am to get it to work.

Added __cplusplus wrapping where necessary to accommodate nvcc
Added GPTLprint_gpustats() call to src/gptl.c to get GPU stats